NBDRWD_PyTorch1.7.1/requirements.txt
opencv-python
matplotlib
tqdm
imageio
scikit-image
scipy
PyTorch 1.7.1
pytorch-wavelets